@@ -1,6 +1,6 @@
 {
   "name": "jiti",
-  "version": "1.20.0",
+  "version": "1.21.0",
   "description": "Runtime typescript and ESM support for Node.js",
   "repository": "unjs/jiti",
   "license": "MIT",
@@ -25,35 +25,35 @@
     "test:bun": "bun --bun test test/bun"
   },
   "devDependencies": {
-    "@babel/core": "^7.22.15",
-    "@babel/plugin-proposal-decorators": "^7.22.15",
-    "@babel/plugin-transform-export-namespace-from": "^7.18.9",
-    "@babel/plugin-transform-nullish-coalescing-operator": "^7.18.6",
-    "@babel/plugin-transform-optional-chaining": "^7.21.0",
+    "@babel/core": "^7.23.2",
+    "@babel/plugin-proposal-decorators": "^7.23.2",
     "@babel/plugin-syntax-class-properties": "^7.12.13",
     "@babel/plugin-syntax-import-assertions": "^7.22.5",
-    "@babel/plugin-transform-modules-commonjs": "^7.22.15",
+    "@babel/plugin-transform-export-namespace-from": "^7.22.11",
+    "@babel/plugin-transform-modules-commonjs": "^7.23.0",
+    "@babel/plugin-transform-nullish-coalescing-operator": "^7.22.11",
+    "@babel/plugin-transform-optional-chaining": "^7.23.0",
     "@babel/plugin-transform-typescript": "^7.22.15",
-    "@babel/preset-typescript": "^7.22.15",
+    "@babel/preset-typescript": "^7.23.2",
     "@babel/template": "^7.22.15",
-    "@babel/types": "^7.22.15",
-    "@types/babel__core": "^7.20.1",
-    "@types/babel__template": "^7.4.1",
-    "@types/node": "^20.5.9",
-    "@types/object-hash": "^3.0.4",
-    "@types/resolve": "^1.20.2",
-    "@types/semver": "^7.5.1",
-    "@vitest/coverage-v8": "^0.34.3",
-    "acorn": "^8.10.0",
+    "@babel/types": "^7.23.0",
+    "@types/babel__core": "^7.20.3",
+    "@types/babel__template": "^7.4.3",
+    "@types/node": "^20.8.9",
+    "@types/object-hash": "^3.0.5",
+    "@types/resolve": "^1.20.4",
+    "@types/semver": "^7.5.4",
+    "@vitest/coverage-v8": "^0.34.6",
+    "acorn": "^8.11.2",
     "babel-plugin-dynamic-import-node": "^2.3.3",
     "babel-plugin-parameter-decorator": "^1.0.16",
     "babel-plugin-transform-typescript-metadata": "^0.3.2",
     "changelogen": "^0.5.5",
     "config": "^3.3.9",
     "create-require": "^1.1.1",
-    "destr": "^2.0.1",
+    "destr": "^2.0.2",
     "escape-string-regexp": "^5.0.0",
-    "eslint": "^8.48.0",
+    "eslint": "^8.52.0",
     "eslint-config-unjs": "^0.2.1",
     "esm": "^3.2.25",
     "estree-walker": "^3.0.3",
@@ -69,13 +69,13 @@
     "semver": "^7.5.4",
     "std-env": "^3.4.3",
     "terser-webpack-plugin": "^5.3.9",
-    "ts-loader": "^9.4.4",
+    "ts-loader": "^9.5.0",
     "tslib": "^2.6.2",
     "typescript": "^5.2.2",
-    "vite": "^4.4.9",
-    "vitest": "^0.34.3",
-    "webpack": "^5.88.2",
+    "vite": "^4.5.0",
+    "vitest": "^0.34.6",
+    "webpack": "^5.89.0",
     "webpack-cli": "^5.1.4"
   },
-  "packageManager": "pnpm@8.7.4"
+  "packageManager": "pnpm@8.8.0"
 }
