@@ -60,104 +104,127 @@
   },
   "homepage": "https://rollupjs.org/",
   "optionalDependencies": {
-    "fsevents": "~2.3.2"
+    "fsevents": "~2.3.2",
+    "@rollup/rollup-darwin-arm64": "4.17.2",
+    "@rollup/rollup-android-arm64": "4.17.2",
+    "@rollup/rollup-win32-arm64-msvc": "4.17.2",
+    "@rollup/rollup-linux-arm64-gnu": "4.17.2",
+    "@rollup/rollup-linux-arm64-musl": "4.17.2",
+    "@rollup/rollup-android-arm-eabi": "4.17.2",
+    "@rollup/rollup-linux-arm-gnueabihf": "4.17.2",
+    "@rollup/rollup-linux-arm-musleabihf": "4.17.2",
+    "@rollup/rollup-win32-ia32-msvc": "4.17.2",
+    "@rollup/rollup-linux-riscv64-gnu": "4.17.2",
+    "@rollup/rollup-linux-powerpc64le-gnu": "4.17.2",
+    "@rollup/rollup-linux-s390x-gnu": "4.17.2",
+    "@rollup/rollup-darwin-x64": "4.17.2",
+    "@rollup/rollup-win32-x64-msvc": "4.17.2",
+    "@rollup/rollup-linux-x64-gnu": "4.17.2",
+    "@rollup/rollup-linux-x64-musl": "4.17.2"
+  },
+  "dependencies": {
+    "@types/estree": "1.0.5"
   },
   "devDependenciesComments": {
-    "@rollup/plugin-typescript": "It appears that 11.1.3 breaks sourcemaps"
+    "@types/node": "Version 18.19.0 breaks chokidar and vite types",
+    "core-js": "We only update manually as every update requires a snapshot update"
   },
   "devDependencies": {
-    "@codemirror/commands": "^6.2.5",
-    "@codemirror/lang-javascript": "^6.2.1",
-    "@codemirror/language": "^6.9.0",
-    "@codemirror/search": "^6.5.3",
-    "@codemirror/state": "^6.2.1",
-    "@codemirror/view": "^6.19.0",
+    "@codemirror/commands": "^6.5.0",
+    "@codemirror/lang-javascript": "^6.2.2",
+    "@codemirror/language": "^6.10.1",
+    "@codemirror/search": "^6.5.6",
+    "@codemirror/state": "^6.4.1",
+    "@codemirror/view": "^6.26.3",
     "@jridgewell/sourcemap-codec": "^1.4.15",
-    "@mermaid-js/mermaid-cli": "^10.4.0",
-    "@rollup/plugin-alias": "^5.0.0",
-    "@rollup/plugin-buble": "^1.0.2",
-    "@rollup/plugin-commonjs": "^25.0.4",
-    "@rollup/plugin-json": "^6.0.0",
-    "@rollup/plugin-node-resolve": "^15.2.1",
-    "@rollup/plugin-replace": "^5.0.2",
-    "@rollup/plugin-terser": "^0.4.3",
-    "@rollup/plugin-typescript": "11.1.2",
-    "@rollup/pluginutils": "^5.0.4",
-    "@types/estree": "1.0.1",
-    "@types/mocha": "^10.0.1",
-    "@types/node": "~14.18.61",
-    "@types/yargs-parser": "^21.0.0",
-    "@typescript-eslint/eslint-plugin": "^6.7.2",
-    "@typescript-eslint/parser": "^6.7.2",
-    "@vue/eslint-config-prettier": "^8.0.0",
-    "@vue/eslint-config-typescript": "^12.0.0",
-    "acorn": "^8.10.0",
+    "@mermaid-js/mermaid-cli": "^10.8.0",
+    "@napi-rs/cli": "^2.18.2",
+    "@rollup/plugin-alias": "^5.1.0",
+    "@rollup/plugin-buble": "^1.0.3",
+    "@rollup/plugin-commonjs": "^25.0.7",
+    "@rollup/plugin-json": "^6.1.0",
+    "@rollup/plugin-node-resolve": "^15.2.3",
+    "@rollup/plugin-replace": "^5.0.5",
+    "@rollup/plugin-terser": "^0.4.4",
+    "@rollup/plugin-typescript": "^11.1.6",
+    "@rollup/pluginutils": "^5.1.0",
+    "@shikijs/vitepress-twoslash": "^1.3.0",
+    "@types/eslint": "^8.56.10",
+    "@types/inquirer": "^9.0.7",
+    "@types/mocha": "^10.0.6",
+    "@types/node": "~18.18.14",
+    "@types/yargs-parser": "^21.0.3",
+    "@typescript-eslint/eslint-plugin": "^7.7.1",
+    "@typescript-eslint/parser": "^7.7.1",
+    "@vue/eslint-config-prettier": "^9.0.0",
+    "@vue/eslint-config-typescript": "^13.0.0",
+    "acorn": "^8.11.3",
     "acorn-import-assertions": "^1.9.0",
-    "acorn-jsx": "^5.3.2",
-    "acorn-walk": "^8.2.0",
     "buble": "^0.20.0",
     "builtin-modules": "^3.3.0",
-    "chokidar": "^3.5.3",
+    "chokidar": "^3.6.0",
     "colorette": "^2.0.20",
-    "concurrently": "^8.2.1",
-    "core-js": "^3.32.2",
+    "concurrently": "^8.2.2",
+    "core-js": "3.36.0",
+    "cross-env": "^7.0.3",
     "date-time": "^4.0.0",
     "es5-shim": "^4.6.7",
     "es6-shim": "^0.35.8",
-    "eslint": "^8.49.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-import": "^2.28.1",
-    "eslint-plugin-prettier": "^5.0.0",
-    "eslint-plugin-unicorn": "^48.0.1",
-    "eslint-plugin-vue": "^9.17.0",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-unicorn": "^52.0.0",
+    "eslint-plugin-vue": "^9.25.0",
     "fixturify": "^3.0.0",
     "flru": "^1.0.2",
-    "fs-extra": "^11.1.1",
+    "fs-extra": "^11.2.0",
     "github-api": "^3.4.0",
-    "hash.js": "^1.1.7",
-    "husky": "^8.0.3",
-    "inquirer": "^9.2.11",
+    "husky": "^9.0.11",
+    "inquirer": "^9.2.19",
     "is-reference": "^3.0.2",
-    "lint-staged": "^14.0.1",
+    "lint-staged": "^15.2.2",
     "locate-character": "^3.0.0",
-    "magic-string": "^0.30.3",
-    "mocha": "^10.2.0",
+    "magic-string": "^0.30.10",
+    "mocha": "^10.4.0",
     "nyc": "^15.1.0",
-    "pinia": "^2.1.6",
-    "prettier": "^3.0.3",
+    "pinia": "^2.1.7",
+    "prettier": "^3.2.5",
     "pretty-bytes": "^6.1.1",
-    "pretty-ms": "^8.0.0",
+    "pretty-ms": "^9.0.0",
     "requirejs": "^2.3.6",
-    "rollup": "^3.29.2",
-    "rollup-plugin-license": "^3.1.0",
+    "rollup": "^4.16.3",
+    "rollup-plugin-license": "^3.3.1",
     "rollup-plugin-string": "^3.0.0",
-    "rollup-plugin-thatworks": "^1.0.4",
-    "semver": "^7.5.4",
+    "semver": "^7.6.0",
     "shx": "^0.3.4",
     "signal-exit": "^4.1.0",
     "source-map": "^0.7.4",
     "source-map-support": "^0.5.21",
-    "systemjs": "^6.14.2",
-    "terser": "^5.19.4",
+    "systemjs": "^6.14.3",
+    "terser": "^5.30.4",
     "tslib": "^2.6.2",
-    "typescript": "^5.2.2",
-    "vite": "^4.4.9",
-    "vitepress": "^1.0.0-rc.14",
-    "vue": "^3.3.4",
+    "typescript": "^5.4.5",
+    "vite": "^5.2.10",
+    "vitepress": "^1.1.3",
+    "vue": "^3.4.24",
+    "wasm-pack": "^0.12.1",
     "weak-napi": "^2.0.2",
     "yargs-parser": "^21.1.1"
   },
   "overrides": {
-    "semver": "^7.5.4"
+    "axios": "^1.6.8",
+    "semver": "^7.6.0"
   },
   "files": [
+    "dist/*.node",
     "dist/**/*.js",
     "dist/*.d.ts",
     "dist/bin/rollup",
     "dist/es/package.json"
   ],
   "engines": {
-    "node": ">=14.18.0",
+    "node": ">=18.0.0",
     "npm": ">=8.0.0"
   },
   "exports": {
