@@ -1,6 +1,6 @@
 {
   "name": "sucrase",
-  "version": "3.34.0",
+  "version": "3.35.0",
   "description": "Super-fast alternative to Babel for when you can target modern JS runtimes",
   "author": "Alan Pierce <alangpierce@gmail.com>",
   "license": "MIT",
@@ -50,7 +50,6 @@
   "devDependencies": {
     "@babel/core": "^7.22.5",
     "@jridgewell/trace-mapping": "^0.3.18",
-    "@types/glob": "^7",
     "@types/mocha": "^9.1.1",
     "@types/mz": "^2.7.4",
     "@types/node": "^20.3.2",
@@ -66,7 +65,7 @@
     "mocha": "^10.2.0",
     "nyc": "^15.1.0",
     "prettier": "^2.8.8",
-    "sucrase": "^3.33.0",
+    "sucrase": "^3.34.0",
     "test262-harness": "^10.0.0",
     "ts-interface-builder": "^0.3.3",
     "typescript": "~5.0"
@@ -74,14 +73,14 @@
   "dependencies": {
     "@jridgewell/gen-mapping": "^0.3.2",
     "commander": "^4.0.0",
-    "glob": "7.1.6",
+    "glob": "^10.3.10",
     "lines-and-columns": "^1.1.6",
     "mz": "^2.7.0",
     "pirates": "^4.0.1",
     "ts-interface-checker": "^0.1.9"
   },
   "engines": {
-    "node": ">=8"
+    "node": ">=16 || 14 >=14.17"
   },
   "resolutions": {
     "**/eshost/socket.io": "4.7.0"
